requirements.txt
Django==6.0.3
djangorestframework==3.16.1
pandas==3.0.2
django-cors-headers==4.9.0